requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
spidev==3.5
Werkzeug==2.0.2
requests==2.26.0
Pillow==8.4.0
